DD-902 - mongo connector updates
- removed checksum2
- switched to use async/await
- upgraded to mongodb 3.2.4
