+++ mongo/package.json
@@ -11,7 +11,7 @@
 	"dependencies": {
 		"aws-sdk": "^2.301.0",
 		"leo-sdk": "^2.0.1",
-		"mongodb": "^2.2.26",
+		"mongodb": "^3.2.4",
 		"leo-connector-common": "^1.4.3"
 	}
 }
